+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v3.4.7 (unreleased)
  - Add String() method to all components of the routing connector
+ - Fix many bugs in the routing connector
 
 ## v3.4.6 (2019-06-11)
  - use append but not copy to clone