+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v3.4.8 (unreleased)
+ - Nothing changed yet
+
 ## v3.4.7 (unreleased)
  - Add String() method to all components of the routing connector
  - Fix many bugs in the routing connector